Time: 2 ms (97.18%) | Memory: 45.7 MB (79.15%) - LeetSync

+++ 637-average-of-levels-in-binary-tree/average-of-levels-in-binary-tree.java
@@ -28,7 +28,7 @@ class Solution {
 
             for(int i = 0; i < size; i++) {
                 TreeNode node = q.poll();
-                sum += node.val; 
+                sum += (double)node.val; 
 
                 if(node.left != null) {
                     q.add(node.left);
